wgml: do code more portable
cleanup code, remove some strange code constructs etc.
use int/unsigned typ where possible for better portability between 32/64-bit

+++ bld/wgml/c/gsfc2d.c
@@ -55,7 +55,7 @@
 /*                                                                         */
 /***************************************************************************/
 
-condcode    scr_c2d( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_c2d( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
