+++ bld/wgml/c/gsfc2x.c
@@ -58,7 +58,7 @@ static unsigned char hex( unsigned char c )
 /*                                                                         */
 /***************************************************************************/
 
-condcode    scr_c2x( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_c2x( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
