+++ bld/wgml/c/gsfd2c.c
@@ -53,7 +53,7 @@
 /***************************************************************************/
 
 
-condcode    scr_d2c( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_d2c( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
