+++ bld/wgml/c/gsfdelst.c
@@ -48,7 +48,7 @@
 /*      &'delstr('abcdef',10) ==> abcdef                                   */
 /***************************************************************************/
 
-condcode    scr_delstr( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_delstr( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
