+++ bld/wgml/c/gsflngth.c
@@ -50,7 +50,7 @@
 /*                                                                         */
 /***************************************************************************/
 
-condcode    scr_length( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_length( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
