+++ bld/wgml/c/gsfmin.c
@@ -51,7 +51,7 @@
 /***************************************************************************/
 
 
-condcode    scr_min( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_min( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
