+++ bld/wgml/c/gsfveclp.c
@@ -85,8 +85,8 @@ static  bool    vec_pos;           // true if &'vecpos, false if &'veclastpos
 /*                                                                         */
 /***************************************************************************/
 
-static  condcode    scr_veclp( parm parms[MAX_FUN_PARMS], size_t parmcount,
-                               char * * result, int32_t ressize )
+static  condcode    scr_veclp( parm parms[MAX_FUN_PARMS], unsigned parmcount,
+                               char **result, unsigned ressize )
 {
     char            *   pneedle;
     char            *   pneedlend;
@@ -184,7 +184,7 @@ static  condcode    scr_veclp( parm parms[MAX_FUN_PARMS], size_t parmcount,
  *
  */
 
-condcode    scr_vecpos( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_vecpos( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     vec_pos = true;
     return( scr_veclp( parms, parmcount, result, ressize ) );
@@ -196,7 +196,7 @@ condcode    scr_vecpos( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * re
  *
  */
 
-condcode    scr_veclastpos( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_veclastpos( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     vec_pos = false;
     return( scr_veclp( parms, parmcount, result, ressize ) );
