+++ bld/wgml/c/gsfwidth.c
@@ -56,7 +56,7 @@
 /*     &'width(AAA,X) ==> invalid 'type' operand                           */
 /***************************************************************************/
 
-condcode    scr_width( parm parms[MAX_FUN_PARMS], size_t parmcount, char * * result, int32_t ressize )
+condcode    scr_width( parm parms[MAX_FUN_PARMS], unsigned parmcount, char **result, unsigned ressize )
 {
     char            *   pval;
     char            *   pend;
